+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * Added `Grav\Framework\Form\FormFlashFile` class which implements `UploadedFileInterface` from PSR-7
     * Added `Grav\Framework\Filesystem\Filesystem` class with methods to manipulate stream URLs
     * Added new `$grav['filesystem']` service using an instance of the new `Filesystem` object
+    * Added `{% render object layout: 'default' with { variable: true } %}` for Flex objects and collections
     * Grav 1.6: Flex: Added support for custom object index classes (API compatibility break)
 1. [](#improved)
     * Improved `Grav\Framework\File\Formatter` classes to have abstract parent class and some useful methods